+++ src/@configs/navigation.js
@@ -34,6 +34,7 @@ import { RiProgress7Line        as LoadingIcon        } from "react-icons/ri";
 import { BsFillMarkdownFill     as MarkdownIcon       } from 'react-icons/bs' ;
 import { GiDominoMask           as MaskIcon           } from "react-icons/gi";
 import { RiLayoutMasonryFill    as MasonryIcon        } from "react-icons/ri";
+import { TbSlashes             as BreadcrumbsIcon    } from "react-icons/tb";
 import { TfiLayoutMenuSeparated as MenuIcon           } from "react-icons/tfi" ;
 import { LuLayoutPanelTop       as MegamenuIcon       } from "react-icons/lu";
 import { SiDialogflow           as ModalIcon          } from "react-icons/si";
@@ -115,9 +116,10 @@ const navigation =
                 Icon : NavigationIcon ,
                 items :
                 [
-                    { id : 'pagination' , type : LINK  , Icon : PaginationIcon , path  : '/lab/pagination' } ,
-                    { id : 'menus'      , type : LINK  , Icon : MenuIcon       , path  : '/lab/menus'      } ,
-                    { id : 'megamenu'   , type : LINK  , Icon : MegamenuIcon   , path  : '/lab/megamenu'   } ,
+                    { id : 'breadcrumbs' , type : LINK  , Icon : BreadcrumbsIcon , path  : '/lab/breadcrumbs' } ,
+                    { id : 'pagination'  , type : LINK  , Icon : PaginationIcon  , path  : '/lab/pagination'  } ,
+                    { id : 'menus'       , type : LINK  , Icon : MenuIcon        , path  : '/lab/menus'       } ,
+                    { id : 'megamenu'    , type : LINK  , Icon : MegamenuIcon    , path  : '/lab/megamenu'    } ,
                 ]
             } ,
             {